Table viewer
Added a table viewer to help with debugging tracking apps. Shows the
contents of the two SQLite tables on a single page. Included in a dev
folder to allow for more developer additions later. If not debugging it
might be an idea to delete this folder to avoid revealing unwanted
information.