Fixed Google Analytics and update problems
Added Google Analytics to the whole system so it should actually work
now. Also fixed minor problem on AJAX call when updating large numbers
of points, was adding them to the map backwards resulting in strange
straight lines, now SQL forces the points to be given in the correct
order.