Emulator, graphs better styles
Added an emulator (in /dev/emulator) to help with developing, pick a
point to start and it will provide random movement so development for
the live updates is easier.

Added graphs that show speed and altitude, this updates live and can be
accessed via the dropdown box below the information panel. The addition
of this means the mapupdate.php file was required to provide some more
information. Hovering over the graph provides a marker on the map to
show corresponding location (markers added to resources folder)

Images have increased resolution and the styling is hopefully better for
mobiles.